requirements.txt
# Core Python
python>=3.8

# Voice Processing
SpeechRecognition==3.10.0
PyAudio==0.2.11
pvporcupine==3.0.0
openai-whisper==20231117
pygame==2.5.0
gTTS==2.3.2
pydub==0.25.1

# AI & ML
rasa==3.6.15
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2

# Web & API
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
requests==2.31.0
python-socketio==5.10.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
prompt-toolkit==3.0.43

# Hardware (Raspberry Pi)
RPi.GPIO==0.7.1
gpiozero==1.6.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0